renaming and moving File upload logic into own package
