+++ system/admin/upload/upload.go
@@ -1,4 +1,4 @@
-package admin
+package upload
 
 import (
 	"fmt"
@@ -10,8 +10,8 @@ import (
 	"time"
 )
 
-// StoreFileUploads stores file uploads at paths like /YYYY/MM/filename.ext
-func StoreFileUploads(req *http.Request) (map[string]string, error) {
+// StoreFiles stores file uploads at paths like /YYYY/MM/filename.ext
+func StoreFiles(req *http.Request) (map[string]string, error) {
 	err := req.ParseMultipartForm(1024 * 1024 * 4) // maxMemory 4MB
 	if err != nil {
 		return nil, fmt.Errorf("%s", err)
